requirements.txt
customtkinter>=5.2.0
darkdetect>=0.8.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.2.0
tensorflow>=2.12.0
Pillow>=9.4.0